clean-up.
Former-commit-id: bc2e9235d506185d24f5fedb393ad2f394d8ab57

+++ src/cell.h
@@ -109,7 +109,7 @@ struct cell {
     float mom[3], ang[3];
     
     /* Potential and kinetic energy of particles in this cell. */
-    float epot, ekin;
+    double epot, ekin;
     
     /* Number of particles updated in this cell. */
     int updated;
